Move internal collections extensions to a separate file.
