+++ coil-base/src/main/java/coil/util/Extensions.kt
@@ -62,78 +62,6 @@ internal fun Bitmap.Config?.getBytesPerPixel(): Int {
     }
 }
 
-/** Functionally the same as [Iterable.forEach] except it generates an index-based loop that doesn't use an [Iterator]. */
-internal inline fun <T> List<T>.forEachIndices(action: (T) -> Unit) {
-    for (i in indices) {
-        action(get(i))
-    }
-}
-
-/** Functionally the same as [Iterable.map] except it generates an index-based loop that doesn't use an [Iterator]. */
-internal inline fun <R, T> List<R>.mapIndices(transform: (R) -> T): List<T> {
-    val destination = ArrayList<T>(size)
-    for (i in indices) {
-        destination += transform(get(i))
-    }
-    return destination
-}
-
-/** Return the first non-null value returned by [transform]. Generate an index-based loop that doesn't use an [Iterator]. */
-internal inline fun <R, T> List<R>.firstNotNullIndices(transform: (R) -> T?): T? {
-    for (i in indices) {
-        transform(get(i))?.let { return it }
-    }
-    return null
-}
-
-/** Functionally the same as [Iterable.find] except it generates an index-based loop that doesn't use an [Iterator]. */
-internal inline fun <T> List<T>.findIndices(predicate: (T) -> Boolean): T? {
-    for (i in indices) {
-        val value = get(i)
-        if (predicate(value)) {
-            return value
-        }
-    }
-    return null
-}
-
-/** Functionally the same as [Iterable.fold] except it generates an index-based loop that doesn't use an [Iterator]. */
-internal inline fun <T, R> List<T>.foldIndices(initial: R, operation: (R, T) -> R): R {
-    var accumulator = initial
-    for (i in indices) {
-        accumulator = operation(accumulator, get(i))
-    }
-    return accumulator
-}
-
-/** Removes values from the list as determined by the [predicate]. */
-internal inline fun <T> MutableList<T>.removeIfIndices(predicate: (T) -> Boolean) {
-    var numDeleted = 0
-
-    for (rawIndex in indices) {
-        val index = rawIndex - numDeleted
-        val value = get(index)
-
-        if (predicate(value)) {
-            removeAt(index)
-            numDeleted++
-        }
-    }
-}
-
-internal inline fun <T> MutableList<T>.removeLast(): T? = if (isNotEmpty()) removeAt(lastIndex) else null
-
-internal inline fun <K, V, R : Any> Map<K, V>.mapNotNullValues(transform: (Map.Entry<K, V>) -> R?): Map<K, R> {
-    val destination = mutableMapOf<K, R>()
-    for (entry in this) {
-        val value = transform(entry)
-        if (value != null) {
-            destination[entry.key] = value
-        }
-    }
-    return destination
-}
-
 internal inline fun ActivityManager.isLowRamDeviceCompat(): Boolean {
     return SDK_INT < 19 || isLowRamDevice
 }
@@ -203,10 +131,6 @@ internal val View.requestManager: ViewTargetRequestManager
         }
         return manager
     }
-
-internal typealias MultiMutableList<R, T> = MutableList<Pair<R, T>>
-
-internal typealias MultiList<R, T> = List<Pair<R, T>>
 
 internal val DataSource.emoji: String
     get() = when (this) {
